requirements-e2e.txt
pytest
pytest-playwright
pytest-json-report
